Use size_t for matrix indecies

+++ src/matrix.cpp
@@ -1,6 +1,6 @@
 #include "matrix.hpp"
 
-Matrix::Matrix(int rows, int columns)
+Matrix::Matrix(size_t rows, size_t columns)
     : rows(rows), cols(columns) {
     data = new double[rows*cols];
 }
@@ -12,8 +12,8 @@ Matrix::Matrix(std::initializer_list<std::initializer_list<double>> values)
     }
 
     data = new double[rows*cols];
-    for(int r=0; r<rows; r++){
-        for(int c=0; c<cols; c++){
+    for(size_t r=0; r<rows; r++){
+        for(size_t c=0; c<cols; c++){
             set(r,c, *((values.begin()+r)->begin()+c) );
         }
     }
@@ -28,23 +28,23 @@ Matrix::Matrix(const Matrix& other)
 }
 Matrix Matrix::T() const{ //transpose
     Matrix nm(cols, rows);
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < cols; c++){
             nm.set(c, r, get(r,c));
         }
     }
     return nm;
 }
 void Matrix::fill(double value){
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < cols; c++){
             set(r, c, value);
         }
     }
 }
 void Matrix::map(std::function<double(double)> transform){
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < cols; c++){
             double value = transform(get(r,c));
             set(r, c, value);
         }
@@ -53,8 +53,8 @@ void Matrix::map(std::function<double(double)> transform){
 std::string Matrix::toString(int precision) const{
     std::stringstream ss;
     ss.precision(precision);
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < cols; c++){
             ss << get(r,c) << "  ";
         }
         ss << "\n";
@@ -64,8 +64,8 @@ std::string Matrix::toString(int precision) const{
 Matrix& Matrix::operator+= (const Matrix& rhs){
     assert (this->rows == rhs.rows);
     assert (this->cols == rhs.cols);
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < cols; c++){
             double value = this->get(r,c) + rhs.get(r,c);
             set(r, c, value);
         }
@@ -76,8 +76,8 @@ Matrix Matrix::operator- (const Matrix& rhs) const{
     assert (this->rows == rhs.rows);
     assert (this->cols == rhs.cols);
     Matrix nm(rows, cols);
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < cols; c++){
             double value = this->get(r,c) - rhs.get(r,c);
             nm.set(r, c, value);
         }
@@ -88,8 +88,8 @@ Matrix Matrix::operator+ (const Matrix& rhs) const{
     assert (this->rows == rhs.rows);
     assert (this->cols == rhs.cols);
     Matrix nm(rows, cols);
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < cols; c++){
             double value = this->get(r,c) + rhs.get(r,c);
             nm.set(r, c, value);
         }
@@ -99,10 +99,10 @@ Matrix Matrix::operator+ (const Matrix& rhs) const{
 Matrix Matrix::operator* (const Matrix& rhs) const{
     assert (this->cols == rhs.rows);
     Matrix nm(rows, rhs.cols);
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < rhs.cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < rhs.cols; c++){
             double value = 0.0;
-            for(int i=0; i<cols; i++){
+            for(size_t i=0; i<cols; i++){
                 value += get(r, i) * rhs.get(i, c);
             }
             nm.set(r, c, value);
@@ -113,13 +113,13 @@ Matrix Matrix::operator* (const Matrix& rhs) const{
 Matrix Matrix::operator| (const Matrix& rhs) const{
     assert (this->rows == rhs.rows);
     Matrix nm(rows, cols+rhs.cols);
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < cols; c++){
             nm.set(r, c, this->get(r,c));
         }
     }
-    for(int r = 0; r < rows; r++){
-        for(int c = 0; c < rhs.cols; c++){
+    for(size_t r = 0; r < rows; r++){
+        for(size_t c = 0; c < rhs.cols; c++){
             nm.set(r, cols+c, rhs.get(r,c));
         }
     }
